Rename main.ts to compiler.ts
Because that's what it really is
